release version 2.1.3

+++ src/main/java/io/github/dataforseo/client/model/BacklinksIdListResultInfo.java
@@ -154,9 +154,9 @@ public class BacklinksIdListResultInfo  {
 
   public static final String SERIALIZED_NAME_COST = "cost";
   @SerializedName(SERIALIZED_NAME_COST)
-  private Double cost;
+  private Integer cost;
 
-  public BacklinksIdListResultInfo cost(Double cost) {
+  public BacklinksIdListResultInfo cost(Integer cost) {
     this.cost = cost;
     return this;
   }
@@ -166,11 +166,11 @@ public class BacklinksIdListResultInfo  {
    * @return cost
    */
   @javax.annotation.Nullable
-  public Double getCost() {
+  public Integer getCost() {
     return cost;
   }
 
-  public void setCost(Double cost) {
+  public void setCost(Integer cost) {
     this.cost = cost;
   }
