+++ src/main/java/io/github/dataforseo/client/model/MerchantErrorsResultInfo.java
@@ -223,9 +223,9 @@ public class MerchantErrorsResultInfo  {
 
   public static final String SERIALIZED_NAME_HTTP_TIME = "http_time";
   @SerializedName(SERIALIZED_NAME_HTTP_TIME)
-  private Double httpTime;
+  private Integer httpTime;
 
-  public MerchantErrorsResultInfo httpTime(Double httpTime) {
+  public MerchantErrorsResultInfo httpTime(Integer httpTime) {
     this.httpTime = httpTime;
     return this;
   }
@@ -235,11 +235,11 @@ public class MerchantErrorsResultInfo  {
    * @return httpTime
    */
   @javax.annotation.Nullable
-  public Double getHttpTime() {
+  public Integer getHttpTime() {
     return httpTime;
   }
 
-  public void setHttpTime(Double httpTime) {
+  public void setHttpTime(Integer httpTime) {
     this.httpTime = httpTime;
   }
