+++ src/main/java/io/github/dataforseo/client/model/SerpGoogleOrganicLiveAdvancedResultInfo.java
@@ -47,7 +47,7 @@ public class SerpGoogleOrganicLiveAdvancedResultInfo  {
   }
 
   /**
-   * 
+   * <em>keyword received in a POST array<br></em><strong>the keyword is returned with decoded %## (plus character '+' will be decoded to a space character)</strong>
    * @return keyword
    */
   @javax.annotation.Nullable
@@ -93,7 +93,7 @@ public class SerpGoogleOrganicLiveAdvancedResultInfo  {
   }
 
   /**
-   * 
+   * <em>search engine domain in a POST array</em>
    * @return seDomain
    */
   @javax.annotation.Nullable
@@ -116,7 +116,7 @@ public class SerpGoogleOrganicLiveAdvancedResultInfo  {
   }
 
   /**
-   * 
+   * <em>location code in a POST array</em>
    * @return locationCode
    */
   @javax.annotation.Nullable
@@ -139,7 +139,7 @@ public class SerpGoogleOrganicLiveAdvancedResultInfo  {
   }
 
   /**
-   * 
+   * <em>language code in a POST array</em>
    * @return languageCode
    */
   @javax.annotation.Nullable
@@ -162,7 +162,7 @@ public class SerpGoogleOrganicLiveAdvancedResultInfo  {
   }
 
   /**
-   * 
+   * <em>direct URL to search engine results</em><br>you can use it to make sure that we provided accurate results
    * @return checkUrl
    */
   @javax.annotation.Nullable
@@ -185,7 +185,7 @@ public class SerpGoogleOrganicLiveAdvancedResultInfo  {
   }
 
   /**
-   * 
+   * <em>date and time when the result was received</em><br>in the UTC format: “yyyy-mm-dd hh-mm-ss +00:00”<br>example:<br><code class='long-string'>2019-11-15 12:57:46 +00:00</code>
    * @return datetime
    */
   @javax.annotation.Nullable
@@ -208,7 +208,7 @@ public class SerpGoogleOrganicLiveAdvancedResultInfo  {
   }
 
   /**
-   * 
+   * <em>autocorrection of the search engine</em><br>if the search engine provided results for a keyword that was corrected, we will specify the keyword corrected by the search engine and the type of autocorrection
    * @return spell
    */
   @javax.annotation.Nullable
@@ -231,7 +231,7 @@ public class SerpGoogleOrganicLiveAdvancedResultInfo  {
   }
 
   /**
-   * 
+   * <em>search refinement chips</em>
    * @return refinementChips
    */
   @javax.annotation.Nullable
@@ -254,7 +254,7 @@ public class SerpGoogleOrganicLiveAdvancedResultInfo  {
   }
 
   /**
-   * 
+   * <em>types of search results in SERP</em><br>contains types of search results (<code>items</code>) found in SERP.<br>possible item types:<br><span><a href='#answer_box'><code>answer_box</code></a>, <a href='#app'><code>app</code></a>, <a href='#carousel'><code>carousel</code></a>, <a href='#multi_carousel'><code>multi_carousel</code></a>, <a href='#featured_snippet'><code>featured_snippet</code></a>, <a href='#google_flights'><code>google_flights</code></a>, <a href='#google_reviews'><code>google_reviews</code></a>, <a href='#third_party_reviews'><code>third_party_reviews</code></a>, <a href='#images'><code>images</code></a>, <a href='#jobs'><code>jobs</code></a>, <a href='#knowledge_graph'><code>knowledge_graph</code></a>, <a href='#local_pack'><code>local_pack</code></a>, <a href='#hotels_pack'><code>hotels_pack</code></a>, <a href='#map'><code>map</code></a>, <a href='#organic'><code>organic</code></a>, <a href='#paid'><code>paid</code></a>, <a href='#people_also_ask'><code>people_also_ask</code></a>, <a href='#related_searches'><code>related_searches</code></a>, <a href='#people_also_search'><code>people_also_search</code></a>, <a href='#shopping'><code>shopping</code></a>, <a href='#top_stories'><code>top_stories</code></a>, <a href='#twitter'><code>twitter</code></a>, <a href='#video'><code>video</code></a>, <a href='#events'><code>events</code></a>, <a href='#recipes'><code>recipes</code></a>, <a href='#top_sights'><code>top_sights</code></a>, <a href='#scholarly_articles'><code>scholarly_articles</code></a>, <a href='#popular_products'><code>popular_products</code></a>,  <a href='#questions_and_answers'><code>questions_and_answers</code></a>, <a href='#find_results_on'><code>find_results_on</code></a>, <a href='#stocks_box'><code>stocks_box</code></a>, <a href='#commercial_units'><code>commercial_units</code></a>, <a href='#local_services'><code>local_services</code></a>, <a href='#google_hotels'><code>google_hotels</code></a>, <a href='#math_solver'><code>math_solver</code></a>, <a href='#currency_box'><code>currency_box</code></a>,<a href='#product_considerations'><code>product_considerations</code></a>, <a href='#short_videos'><code>short_videos</code></a>, <a href='#refine_products'><code>refine_products</code></a>, <a href='#perspectives'><code>perspectives</code></a>, <a href='#discussions_and_forums'><code>discussions_and_forums</code></a></span>, <a href='#compare_sites'><code>compare_sites</code></a></span>, <a href='#ai_overview'><code>ai_overview</code></a></span>
    * @return itemTypes
    */
   @javax.annotation.Nullable
@@ -269,69 +269,69 @@ public class SerpGoogleOrganicLiveAdvancedResultInfo  {
 
   public static final String SERIALIZED_NAME_SE_RESULTS_COUNT = "se_results_count";
   @SerializedName(SERIALIZED_NAME_SE_RESULTS_COUNT)
-  private Integer seResultsCount;
+  private Long seResultsCount;
 
-  public SerpGoogleOrganicLiveAdvancedResultInfo seResultsCount(Integer seResultsCount) {
+  public SerpGoogleOrganicLiveAdvancedResultInfo seResultsCount(Long seResultsCount) {
     this.seResultsCount = seResultsCount;
     return this;
   }
 
   /**
-   * 
+   * <em> total number of results in SERP</em>
    * @return seResultsCount
    */
   @javax.annotation.Nullable
-  public Integer getSeResultsCount() {
+  public Long getSeResultsCount() {
     return seResultsCount;
   }
 
-  public void setSeResultsCount(Integer seResultsCount) {
+  public void setSeResultsCount(Long seResultsCount) {
     this.seResultsCount = seResultsCount;
   }
 
 
   public static final String SERIALIZED_NAME_PAGES_COUNT = "pages_count";
   @SerializedName(SERIALIZED_NAME_PAGES_COUNT)
-  private Integer pagesCount;
+  private Long pagesCount;
 
-  public SerpGoogleOrganicLiveAdvancedResultInfo pagesCount(Integer pagesCount) {
+  public SerpGoogleOrganicLiveAdvancedResultInfo pagesCount(Long pagesCount) {
     this.pagesCount = pagesCount;
     return this;
   }
 
   /**
-   * 
+   * <em>total search results pages retrieved</em><br>total number of retrieved SERPs in the result
    * @return pagesCount
    */
   @javax.annotation.Nullable
-  public Integer getPagesCount() {
+  public Long getPagesCount() {
     return pagesCount;
   }
 
-  public void setPagesCount(Integer pagesCount) {
+  public void setPagesCount(Long pagesCount) {
     this.pagesCount = pagesCount;
   }
 
 
   public static final String SERIALIZED_NAME_ITEMS_COUNT = "items_count";
   @SerializedName(SERIALIZED_NAME_ITEMS_COUNT)
-  private Integer itemsCount;
+  private Long itemsCount;
 
-  public SerpGoogleOrganicLiveAdvancedResultInfo itemsCount(Integer itemsCount) {
+  public SerpGoogleOrganicLiveAdvancedResultInfo itemsCount(Long itemsCount) {
     this.itemsCount = itemsCount;
     return this;
   }
 
   /**
-   * 
+   * <em>the number of results returned in the <strong><code>items</code></strong> array</em>
    * @return itemsCount
    */
   @javax.annotation.Nullable
-  public Integer getItemsCount() {
+  public Long getItemsCount() {
     return itemsCount;
   }
 
-  public void setItemsCount(Integer itemsCount) {
+  public void setItemsCount(Long itemsCount) {
     this.itemsCount = itemsCount;
   }
 
@@ -346,7 +346,7 @@ public class SerpGoogleOrganicLiveAdvancedResultInfo  {
   }
 
   /**
-   * 
+   * <em>items of the element</em>
    * @return items
    */
   @javax.annotation.Nullable
